Adding main PyBank python file (#5)
* Adding main PyBank python file

This will serve as the primary script to run for the PyBank problem. At the time of publishing, it is only able to read in the file from a local directory that has not yet been synced to this repository. This will be changed in the future and the file path will be updated to reflect that.

* Add files via upload

This will serve as the primary script to run for the PyPoll problem. At the time of publishing, it is only able to read in the file from a local directory that has not yet been synced to this repository. This will be changed in the future and the file path will be updated to reflect that.