Sy flores 11 export file (#26)
* Update main.py

Added code to export the election_results string to a the election_results.txt file. Also updated the csv file path to run on elections.csv. It was previously being run on a local elections_full.csv file as it contained the complete data set.

* Add files via upload

Uploaded the file created by the main.py script. This contains the elections results.